Wybór okna dla filtrów

+++ app/Discrete/ImpulseNoise.py
@@ -14,7 +14,7 @@ class ImpulseNoise(DiscreteSignal):
         self.p = p / 100
 
     def generate_data(self):
-        for i in range(self.d * self.f):
+        for i in range(int(self.d * self.f)):
             t = self.t1 + i / self.f
             self.data.append(self.calculate_data(t))
             self.indexes.append(t)
